Say what a javadoc is about before saying what its parts are
A comment opening on `@param`, `@throws` or `@see` publishes no summary, so
the reader of the generated page is told the parts of an answer without being
told what the answer is. Each one now says it, and the three links that
resolved to nothing name what they meant.

Beside that, the findings that were dead or ambiguous rather than unwritten: a
list built and never read, an index removal spelled as a value removal, an
absolute value that stays negative at the least int, an enum compared by its
ordinal, and a nested type reached by a name it is not declared under.

Two files carry a grouping from the commit before this one; their other change
belongs here.

+++ souther-compiler/src/test/java/souther/compiler/numeric/WhatTheDifferenceBoundsSayIsWhatThePointsSayTest.java
@@ -181,7 +181,7 @@ class WhatTheDifferenceBoundsSayIsWhatThePointsSayTest {
         int howMany = 2 + dice.nextInt(4);
         for (int i = 0; i < howMany; i++) {
             String one = POSITIONS.get(dice.nextInt(POSITIONS.size()));
-            long threshold = LOW + dice.nextInt(HIGH - LOW + 1);
+            long threshold = LOW + (long) dice.nextInt(HIGH - LOW + 1);
             boolean strict = dice.nextBoolean();
             if (dice.nextBoolean()) {
                 Rel rel = dice.nextBoolean()
